refactor: relativeTimeText 挪回 profile，去掉冗余 toLong
两处收尾：

- relativeTimeText 从 ui/common 挪回 GithubProfileScreen 作 private。当初提取是为了
  让列表的抓取时机行复用它，但时机行最终走的是日期锚定（DateTimeUtils.updateStamp），
  一次都没用上——单一调用点的东西放在 common 里只是无谓的公共 API 扩散。
  「去掉 remember」那个修复保留：缓存住 now 会让「3 小时前」在后台切回后仍显示旧值。
- updateStamp 里 toEpochDays() 在本项目的 kotlinx-datetime 版本已返回 Long，
  两处 .toLong() 是冗余转换（IDE 有提示）。

+++ shared/src/commonMain/kotlin/whl/trending/ai/core/DateTimeUtils.kt
@@ -66,7 +66,7 @@ object DateTimeUtils {
         val local = instant.toLocalDateTime(tz)
         val today = now.toLocalDateTime(tz).date
         val time = "${local.hour.toString().padStart(2, '0')}:${local.minute.toString().padStart(2, '0')}"
-        val diffDays = today.toEpochDays().toLong() - local.date.toEpochDays().toLong()
+        val diffDays = today.toEpochDays() - local.date.toEpochDays()
         val unit = when {
             diffDays <= 0L -> UpdateStampUnit.TODAY
             diffDays == 1L -> UpdateStampUnit.YESTERDAY
